Get brick/vty working with pacman plus lots of ghost code
This is a brick/vty version of the code with a moving pacman and code to
move the ghosts, but the code is not tied in.  However, it has a serious
performance issue as, for every 'tick' the brick code re-evaluates the
entire screen to see if it has changed, even if we know nothing has
moved.  This is very inefficient and results in lots of CPU usage.  It's
effectively not workable.  At this point the code is going to change,
and an ncurses version is going to be developed on a separate branch.
This is likely to be the last verion that uses brick/vty.