Switch from Brick to ncurses and lots of other changes
The switch from Brick/Vty to ncurses was purely triggered for
performance reasons.  Pac-man needs a tick, where different things
happen based on varying numbers of ticks having passed.  In Brick this
means that it tried to work out if it had to update the display 60 times
per second which meant it consumed lots of CPU time doing diffs it
didn't need to do.  Due to the architecture of Brick, it wasn't
obvious/easy to work out how to avoid the 'diffing' without pulling out
large parts of Brick into the main program.  Thus, the switch to
ncurses, which whilst not as elegant, means whether to update the
display can be finely controlled.  This is done via the DrawList in the
pacman.hs module.

The other changes are around adding more ghost functionality (but not
actually tying it in yet) and coping with the changes caused by moving
off Brick.  This isn't completely done as the keyboard handling hasn't
been added - that's for the next commit.

This version compiles and puts the initial display up.