Remove custom DrawPacman, DrawGhost and use DrawGridAt
This removes the custom draw types for pacman and ghost, and bulks out
the DrawGridAt type handling so that it draws the Pacman and Ghosts.
Now we can easily reduce the drawlist down using nub (which should be
okay as the list is likely to be small) and then do the draw calls.